All Nav expect Bar done
