+++ app/src/main/java/com/example/espproject/AddIngredientDialog.java
@@ -11,28 +11,8 @@ import android.widget.Button;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
-import androidx.navigation.fragment.NavHostFragment;
-
-import com.example.espproject.databinding.DialogAddIngredientBinding;
 
 public class AddIngredientDialog extends DialogFragment {
-
-    private DialogAddIngredientBinding binding;
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        binding = DialogAddIngredientBinding.inflate(inflater, container, false);
-        return binding.getRoot();
-    }
-    public void onViewCreated(@NonNull View view, Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        binding.buttonAdditem.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                NavHostFragment.findNavController(AddIngredientDialog.this)
-                        .navigate(R.id.action_addIngredientDialog_to_pantryFragment);
-            }
-        });
-    }
 
     // callback interface for passing data back to PantryActivity
     public interface AddItem {
