+++ app/src/main/java/com/example/espproject/PantryUsersDialog.java
@@ -11,17 +11,11 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 
-import com.example.espproject.databinding.DialogPantryUsersBinding;
+
 
 public class PantryUsersDialog extends DialogFragment {
 
     // TODO: declare UI components as fields
-    private DialogPantryUsersBinding binding;
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        binding = DialogPantryUsersBinding.inflate(inflater, container, false);
-        return binding.getRoot();
-    }
 
     @NonNull
     @Override
