+++ app/src/main/res/layout/activity_main.xml
@@ -18,4 +18,12 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:navGraph="@navigation/mynav" />
+
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:layout_width="416dp"
+        android:layout_height="104dp"
+        android:background="#5B098E"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
